# docs/openapi.courtly.yaml
openapi: 3.1.0
info:
  title: Courtly API
  version: 1.0.0
  description: >
    Full-scope API for Courtly tennis court booking platform.
    Includes auth, courts, day/week availability, hold-based booking,
    user cabinet, moderator/admin dashboard, dynamic RBAC/policy management,
    and event log operations.
servers:
  - url: https://api.courtly.local
    description: Production
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/courts/{court_id}:
    get:
      tags: [Courts]
      summary: Get court detail
      security: []
      parameters:
        - $ref: "#/components/parameters/CourtId"
      responses:
        "200":
          description: Court detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourtDetail"
        "404":
          description: Court not found
    patch:
      tags: [Courts]
      summary: Update court
      x-rbac:
        permission: courts:update
        ownership: own_court_or_admin
      parameters:
        - $ref: "#/components/parameters/CourtId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourtUpdateRequest"
      responses:
        "200":
          description: Court updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourtDetail"
        "403":
          description: Forbidden
    delete:
      tags: [Courts]
      summary: Delete court
      x-rbac:
        permission: courts:delete
        ownership: own_court_or_admin
      parameters:
        - $ref: "#/components/parameters/CourtId"
      responses:
        "204":
          description: Court deleted
        "403":
          description: Forbidden
...
  /api/me/bookings:
    get:
      tags: [Cabinet]
      summary: List current user bookings
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, past, cancelled, all]
      responses:
        "200":
          description: User bookings
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/BookingResponse"
  /api/me/profile:
    get:
      tags: [Cabinet]
      summary: Get own profile
      responses:
        "200":
          description: Own profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileSelf"
    patch:
      tags: [Cabinet]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileSelf"
  /api/me/favorites:
    get:
      tags: [Cabinet]
      summary: List own favorite courts
      responses:
        "200":
          description: Favorite courts
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/CourtSummary"
    post:
      tags: [Cabinet]
      summary: Add favorite court
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [court_id]
              properties:
                court_id:
                  type: string
                  format: uuid
      responses:
        "204":
          description: Added
  /api/me/reviews:
    post:
      tags: [Cabinet]
      summary: Create review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResponse"
  /api/dashboard/courts/{court_id}/transfer-ownership:
    post:
      tags: [Dashboard]
      summary: Transfer court ownership by email (anti-enumeration)
      x-rbac:
        permission: courts:transfer_ownership
        ownership: own_court_or_admin
      parameters:
        - $ref: "#/components/parameters/CourtId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_email]
              properties:
                target_email:
                  type: string
                  format: email
      responses:
        "202":
          description: Accepted (generic response regardless of account existence)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "If eligible, transfer instructions will be processed."
  /api/dashboard/notifications/email:
    post:
      tags: [Dashboard]
      summary: Send notification email
      x-rbac:
        permission: notifications:send_email
        ownership: policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailNotificationRequest"
      responses:
        "202":
          description: Accepted for processing
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CourtId:
      in: path
      name: court_id
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CourtUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        name: { type: string }
        city: { type: string }
        district: { type: string }
        address: { type: string }
        indoor: { type: boolean }
        surface: { type: string }
        price_per_hour: { type: number, format: float, minimum: 0 }
    CourtSummary:
      type: object
      required: [id, name, city, district, next_available_slot]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        city: { type: string }
        district: { type: string }
        address: { type: string }
        rating: { type: number, format: float }
        distance_km: { type: number, format: float }
        price_from: { type: number, format: float }
        next_available_slot: { type: string, format: date-time }
    CourtDetail:
      allOf:
        - $ref: "#/components/schemas/CourtSummary"
        - type: object
          properties:
            description: { type: string }
            indoor: { type: boolean }
            surface: { type: string }
            media_urls:
              type: array
              items: { type: string, format: uri }
            working_hours:
              type: array
              items:
                $ref: "#/components/schemas/WorkingHours"
    WorkingHours:
      type: object
      required: [weekday, open_time, close_time]
      properties:
        weekday:
          type: integer
          minimum: 0
          maximum: 6
        open_time:
          type: string
          pattern: "^[0-2][0-9]:[0-5][0-9]$"
        close_time:
          type: string
          pattern: "^[0-2][0-9]:[0-5][0-9]$"
...
    BookingResponse:
      type: object
      required: [id, court_id, user_id, status, start_at, end_at, total_price]
      properties:
        id: { type: string, format: uuid }
        court_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        status:
          type: string
          enum: [confirmed, active, completed, cancelled]
        start_at: { type: string, format: date-time }
        end_at: { type: string, format: date-time }
        total_price: { type: number, format: float }
        created_at: { type: string, format: date-time }
    ProfileSelf:
      type: object
      required: [user_id, full_name, email, phone, role]
      properties:
        user_id: { type: string, format: uuid }
        full_name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        role: { type: string }
        created_at: { type: string, format: date-time }
    ProfileUpdateRequest:
      type: object
      properties:
        full_name: { type: string }
        phone: { type: string }
    ReviewCreateRequest:
      type: object
      required: [court_id, rating, text]
      properties:
        court_id: { type: string, format: uuid }
        rating: { type: integer, minimum: 1, maximum: 5 }
        text: { type: string, minLength: 3, maxLength: 1000 }
    ReviewResponse:
      type: object
      required: [id, court_id, rating, text, created_at]
      properties:
        id: { type: string, format: uuid }
        court_id: { type: string, format: uuid }
        rating: { type: integer }
        text: { type: string }
        created_at: { type: string, format: date-time }
...
    EmailNotificationRequest:
      type: object
      required: [subject, body, recipient_scope]
      properties:
        subject: { type: string, minLength: 3, maxLength: 120 }
        body: { type: string, minLength: 3, maxLength: 5000 }
        recipient_scope:
          type: string
          enum: [all_users, active_users, custom_segment]